Moved to json_database, added verbose make
Moved the JSON builder project to a folder json_database so that it is obviously seperate from main program. Updated makefile to reflect this (used cd in make rather than modifying code), and also added a 'verbose' switch to select whether output should be displayed on screen or piped to log.